Add to_vector tests
